Renamed test classes to be consistent. Cleaned up a bit.
