+++ Source/ReferenceTests/Language/ObjectMemberTests.cs
@@ -1,4 +1,4 @@
-using System;
+﻿using System;
 using NUnit.Framework;
 using System.Management.Automation;
 using TestPSSnapIn;
@@ -8,31 +8,8 @@ using System.Collections.Generic;
 namespace ReferenceTests
 {
     [TestFixture]
-    public class ObjectCommandTests : ReferenceTestBase
+    public class ObjectMemberTests : ReferenceTestBase
     {
-        [Test]
-        public void NewObjectCanCreatePSObject()
-        {
-            var results = ReferenceHost.RawExecute("New-Object -Type PSObject");
-            Assert.AreEqual(1, results.Count, "No results");
-            var obj = results[0].BaseObject;
-            Assert.True(obj is PSCustomObject);
-        }
-
-        [Test]
-        public void AddMemberCanAddNoteProperties()
-        {
-            var results = ReferenceHost.RawExecute(NewlineJoin(
-                "$a = New-Object -Type PSObject",
-                "$a | Add-Member -Type NoteProperty -Name TestName -Value TestValue",
-                "$a"
-            ));
-            Assert.AreEqual(1, results.Count, "No results");
-            var obj = results[0];
-            Assert.NotNull(obj.Members["TestName"]);
-            Assert.NotNull(obj.Properties["TestName"]);
-        }
-
         [Test]
         public void CustomPSObjectPropertiesCanBeAccessedCaseInsensitive()
         {
@@ -52,33 +29,6 @@ namespace ReferenceTests
                 "$a.testname"
             ));
             Assert.AreEqual(NewlineJoin(), result);
-        }
-
-        [Test]
-        public void CreatePSObjectWithPropertiesByHashtable()
-        {
-            var result = ReferenceHost.Execute(NewlineJoin(new string[] {
-                "$obj = new-object psobject -property @{foo='abc'; bar='def'}",
-                "$obj.FoO",
-                "$obj.bAR"
-            }));
-            var expected = NewlineJoin(new string[] { "abc", "def" });
-            Assert.AreEqual(expected, result);
-        }
-
-        [Test]
-        public void CreatePSObjectWithPropertiesAreRawObjects()
-        {
-            var expected = new Dictionary<string, object>() {{"foo", "abc"}, {"bar", 3}};
-            var results = ReferenceHost.RawExecute(NewlineJoin(new string[] {
-                "new-object psobject -property @{foo='abc'; bar=3}"
-            }));
-            Assert.AreEqual(1, results.Count);
-            var res = results[0];
-            foreach (var key in expected.Keys)
-            {
-                Assert.AreEqual(expected[key], res.Properties[key].Value);
-            }
         }
 
         [Test]
