+++ Source/ReferenceTests/ReferenceTests.csproj
@@ -44,7 +44,9 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Commands\AddMemberTests.cs" />
     <Compile Include="Commands\AddTypeTests.cs" />
+    <Compile Include="Language\ObjectMemberTests.cs" />
     <Compile Include="Language\Operators\BitwiseTests.cs" />
     <Compile Include="Language\CmdletParameterTests.cs" />
     <Compile Include="API\DefaultScriptScopeTests.cs" />
@@ -56,21 +58,10 @@
     <Compile Include="ReferenceTestBase.cs" />
     <Compile Include="Language\ExitTests.cs" />
     <Compile Include="ReferenceTestInfo.cs" />
-    <Compile Include="ReturnTests.cs" />
-    <Compile Include="PSObjectTests.cs" />
-    <Compile Include="ObjectCommandTests.cs" />
-    <Compile Include="ArrayTests.cs" />
-    <Compile Include="TypeConverterTests.cs" />
-    <Compile Include="GetMemberCommandTests.cs" />
-    <Compile Include="SelectObjectCommandTests.cs" />
-    <Compile Include="CsvCommandTests.cs" />
-    <Compile Include="StringTests.cs" />
-    <Compile Include="FileSystemProviderTests.cs" />
-    <Compile Include="OutFileTests.cs" />
     <Compile Include="WildcardPatternTests.cs" />
     <Compile Include="Language\ReturnTests.cs" />
     <Compile Include="API\PSObjectTests.cs" />
-    <Compile Include="Commands\ObjectTests.cs" />
+    <Compile Include="Commands\NewObjectTests.cs" />
     <Compile Include="Language\ArrayTests.cs" />
     <Compile Include="API\TypeConverterTests.cs" />
     <Compile Include="Commands\GetMemberTests.cs" />
